# Refactoring + Speed improvements