--- a/pkg/processing/top_level_objects.go
+++ b/pkg/processing/top_level_objects.go
@@ -9,13 +9,14 @@ import (
 
 var fileTopLevelObjectsCache = make(map[string][]*ast.DesugaredObject)
 
-func findTopLevelObjectsInFile(vm *jsonnet.VM, filename string) []*ast.DesugaredObject {
-	if _, ok := fileTopLevelObjectsCache[filename]; !ok {
-		rootNode, _, _ := vm.ImportAST("", filename)
-		fileTopLevelObjectsCache[filename] = findTopLevelObjects(nodestack.NewNodeStack(rootNode), vm)
+func findTopLevelObjectsInFile(vm *jsonnet.VM, filename, from string) []*ast.DesugaredObject {
+	key := from + ":" + filename
+	if _, ok := fileTopLevelObjectsCache[key]; !ok {
+		rootNode, _, _ := vm.ImportAST(from, filename)
+		fileTopLevelObjectsCache[key] = findTopLevelObjects(nodestack.NewNodeStack(rootNode), vm)
 	}
 
-	return fileTopLevelObjectsCache[filename]
+	return fileTopLevelObjectsCache[key]
 }
 
 // Find all ast.DesugaredObject's from NodeStack