--- a/pkg/server/execute.go
+++ b/pkg/server/execute.go
@@ -60,7 +60,7 @@ func (s *server) evalItem(ctx context.Context, params *protocol.ExecuteCommandPa
 	log.Infof("fileName: %s", fileName)
 	log.Infof("position: %+v", p)
 
-	_, node := stack.Pop()
+	node := stack.Pop()
 
 	return nil, fmt.Errorf("%v: %+v", reflect.TypeOf(node), node)
 }