Removed unnecessary return values in node stack
Added an import cache bringing some performance improvements

Ran some benchmarks. Pretty good improvement 😄 
```
name                                                             old time/op  new time/op  delta
Definition/goto_deeply_nested_imported_attribute-8                162µs ±13%    12µs ± 6%  -92.57%  (p=0.000 n=20+20)
Definition/goto_deeply_nested_imported_attribute_through_self-8   166µs ±10%    48µs ± 9%  -71.35%  (p=0.000 n=20+20)
```